cleanup client
