+++ src/client/mod.rs
@@ -1,5 +1,2 @@
-pub(crate) mod connection;
-pub(crate) mod init;
-pub(crate) mod input;
-pub(crate) mod movement;
-pub(crate) mod network;
+pub mod network;
+pub mod render;
