+++ src/client/network/init.rs
@@ -3,14 +3,13 @@ use bevy::prelude::*;
 use bevy_renet::{netcode::*, renet::*, RenetClientPlugin};
 use std::{net::UdpSocket, time::SystemTime};
 
-pub struct ClientInitPlug;
+struct ClientInitPlug;
 
 impl Plugin for ClientInitPlug {
     fn build(&self, app: &mut App) {
         let (client, transport) = new_renet_client();
         app.insert_resource(client);
         app.insert_resource(transport);
-        app.add_systems(Startup, setup_camera);
         app.add_plugins(RenetClientPlugin);
         app.add_plugins(NetcodeClientPlugin);
     }
@@ -19,7 +18,6 @@ impl Plugin for ClientInitPlug {
 pub(crate) fn new_renet_client() -> (RenetClient, NetcodeClientTransport) {
     let server_addr = SERVER_ADDR.parse().unwrap();
     let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
-
     let current_time = SystemTime::now()
         .duration_since(SystemTime::UNIX_EPOCH)
         .unwrap();
@@ -32,10 +30,5 @@ pub(crate) fn new_renet_client() -> (RenetClient, NetcodeClientTransport) {
     };
     let transport = NetcodeClientTransport::new(current_time, authentication, socket).unwrap();
     let client = RenetClient::new(ConnectionConfig::default());
-
     (client, transport)
-}
-
-pub(crate) fn setup_camera(mut commands: Commands) {
-    commands.spawn(Camera2d {});
 }
