+++ src/client/network/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod init;
+pub(crate) mod network;
+pub(crate) mod sending;
