+++ src/client/render/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod connection;
+pub(crate) mod init;
+pub(crate) mod input;
+pub(crate) mod movement;
